Throw together a prototype.
This is a basic prototype that shows a login page and shows a
registration page. When registering, it talks to an API at a hard-coded
address, using the raw net/http library.

This is bad. Nobody should use this. It was just the minimum viable
option for testing out the basic idea of this package.